refactor: package rename step3 -> racingcar
3단계 ~ 5단계는 하나의 패키지에서 과제를 수행하므로 패키지명을 더 적절하게 수정.
